requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
django-crispy-forms==1.14.0
idna==3.4
Pillow==9.3.0
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.12
